feat: Websocket data connection updated (#12)
* feat: Enable browser history navigation and align data types

- Enable browser URL integration for Web (JS/Wasm) by implementing `bindToBrowserNavigation` in `WebNavigation.kt`, leveraging the experimental API.
- Standardize actuator and sensor data types from `Int` to `Double` across the application for consistency with the API. This affects:
  - `GreenhouseMessage` and `GreenhouseData` data models.
  - `GreenhouseViewModel` methods (`updateSector`, `updateExtractor`).
  - `HomeScreen` UI components (`ActuatorControls`, `SectorSlider`) and their callbacks.
- Add `timestamp` and `greenhouseId` fields to the `GreenhouseMessage` data model.
- Set default credentials in `LoginScreen.kt` for easier development and testing.

* Apply code review feedback: remove outdated TODOs and fix type documentation (#13)

* Initial plan

* fix: Apply code review feedback from PR #12


---------


---------

+++ composeApp/src/commonMain/kotlin/com/apptolast/greenhousefronts/presentation/viewmodel/GreenhouseViewModel.kt
@@ -87,9 +87,9 @@ class GreenhouseViewModel(
      *
      * @param greenhouseId The greenhouse ID (1, 2, or 3)
      * @param sectorIndex The sector index (0-3 for sectors 1-4)
-     * @param value The new value for the sector (0-100)
+     * @param value The new value for the sector (Double, 0-100)
      */
-    fun updateSector(greenhouseId: Int, sectorIndex: Int, value: Int) {
+    fun updateSector(greenhouseId: Int, sectorIndex: Int, value: Double) {
         viewModelScope.launch {
             _uiState.value = _uiState.value.copy(isLoading = true, error = null, publishSuccess = false)
 
@@ -146,9 +146,9 @@ class GreenhouseViewModel(
      * Updates the extractor (ventilation) value for a specific greenhouse
      *
      * @param greenhouseId The greenhouse ID (1, 2, or 3)
-     * @param value The new value for the extractor (0 or 1 typically)
+     * @param value The new value for the extractor (0.0 or 1.0 typically)
      */
-    fun updateExtractor(greenhouseId: Int, value: Int) {
+    fun updateExtractor(greenhouseId: Int, value: Double) {
         viewModelScope.launch {
             _uiState.value =
                 _uiState.value.copy(isLoading = true, error = null, publishSuccess = false)
